netron-backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pyserial==3.5
python-socketio==5.8.0
flask-socketio==5.3.6
eventlet==0.33.3
